test: consolidate unit test projects
