+++ SlnxMermaid.slnx
@@ -2,8 +2,6 @@
   <Project Path="src/SlnxMermaid.CLI/SlnxMermaid.CLI.csproj" />
   <Project Path="src/SlnxMermaid.Configuration/SlnxMermaid.Configuration.csproj" />
   <Project Path="src/SlnxMermaid.Gui.Avalonia/SlnxMermaid.Gui.Avalonia.csproj" />
-  <Project Path="tests/SlnxMermaid.Gui.Avalonia.Tests/SlnxMermaid.Gui.Avalonia.Tests.csproj" />
-  <Project Path="tests/SlnxMermaid.Configuration.Tests/SlnxMermaid.Configuration.Tests.csproj" />
   <Project Path="src/SlnxMermaid.Core/SlnxMermaid.Core.csproj" Id="d5badfcd-044d-4ead-b30a-699167eac31b" />
   <Project Path="tests/SlnxMermaid.UnitTests/SlnxMermaid.UnitTests.csproj" Id="b6bc326c-ac66-4d57-b144-8e7a66f968f8" />
   <Project Path="tools/SlnxMermaid.SchemaGenerator/SlnxMermaid.SchemaGenerator.csproj" />
