--- tests/SlnxMermaid.Gui.Avalonia.Tests/SlnxMermaid.Gui.Avalonia.Tests.csproj
@@ -1,22 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net10.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <LangVersion>latest</LangVersion>
-    <IsPackable>false</IsPackable>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="coverlet.collector" Version="6.0.4" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="18.0.1" />
-    <PackageReference Include="xunit" Version="2.9.3" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="3.1.5" />
-  </ItemGroup>
-  <ItemGroup>
-    <Using Include="Xunit" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\SlnxMermaid.Configuration\SlnxMermaid.Configuration.csproj" />
-    <ProjectReference Include="..\..\src\SlnxMermaid.Gui.Avalonia\SlnxMermaid.Gui.Avalonia.csproj" />
-  </ItemGroup>
-</Project>
