+++ tests/SlnxMermaid.UnitTests/SlnxMermaid.UnitTests.csproj
@@ -18,6 +18,7 @@
       <IncludeAssets>compile; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="18.0.1" />
+    <PackageReference Include="NJsonSchema" Version="11.6.1" />
     <PackageReference Include="xunit" Version="2.9.3" />
 	  <PackageReference Include="xunit.runner.visualstudio" Version="3.1.5">
 		  <PrivateAssets>all</PrivateAssets>
@@ -29,6 +30,7 @@
     <ProjectReference Include="..\..\src\SlnxMermaid.CLI\SlnxMermaid.CLI.csproj" />
     <ProjectReference Include="..\..\src\SlnxMermaid.Core\SlnxMermaid.Core.csproj" />
     <ProjectReference Include="..\..\src\SlnxMermaid.Configuration\SlnxMermaid.Configuration.csproj" />
+    <ProjectReference Include="..\..\src\SlnxMermaid.Gui.Avalonia\SlnxMermaid.Gui.Avalonia.csproj" />
   </ItemGroup>
 
   <ItemGroup>
